Fix code-style and simplify expressions in views
A lot of the changes here were to convert from spaces to tabs for
consistency-sake. There were several changes away from the alternate
bracing syntax for `if`, `while` and `foreach` in cases where they were
not being used to interleave HTML. Braced syntax should be preferred.

There were a few difficult-to-read examples of intermixed HTML with PHP
that were cleaned to use double-quotes and evaluation of variables
inline rather than concatenation of strings.