Add build script for plugin
This script uses `git-archive` to export an archive of the plugin
that excludes a few files and directories that WordPress does not
need. This includes the SCSS assets, .gitignore, .gitattributes
and eventually could include things like the test-suite and more.

This is meant to provide an easy way of building the archive to be
uploaded as part of the install process in a case where a manual
deployment / install is done outside of Cthulhu.