Modify .gitignore
Adds a few generic ignore rules for things like hidden and temporary
files in unix and OSX, as well as some rules for ignoring annoying
Windows generated files like thumbnails.

Modifies the generated content section to ignore generated CSS, and
adds the phpunit.xml to the repo since it works out of the box,
and shouldn't really require any changes.